requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
requests
pandas
watchdog
dash
dash-bootstrap-components
plotly
python-jose